Move URL detection into lib/url and reuse existing functions for adding and removing URL protocol
